requirements.txt
Django>=4.2
psycopg2-binary
redis
django-redis
celery
cryptography
qrcode[pil]
pytest
pytest-django
python-dotenv
gunicorn
